Renamed the Repository classes of the AzureStorage project to remove the "AzureStorage" prefix.
