+++ Code/Tardigrade.Framework/Tardigrade.Framework.AzureStorage/Tardigrade.Framework.AzureStorage.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFrameworks>netstandard2.1;netstandard2.0;net472;net461</TargetFrameworks>
-    <Version>6.1.2</Version>
+    <Version>7.0.0</Version>
     <Authors>Rafidzal Rafiq</Authors>
     <Product>Tardigrade Framework Library (Azure Storage Integration)</Product>
     <Description>Software framework for supporting coding best practices.</Description>
