+++ Code/Tardigrade.Framework/Tardigrade.Framework.EntityFramework/Tardigrade.Framework.EntityFramework.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFrameworks>net472;net461</TargetFrameworks>
-    <Version>10.0.1</Version>
+    <Version>10.1.0</Version>
     <Authors>Rafidzal Rafiq</Authors>
     <Product>Tardigrade Framework Library (Entity Framework Integration)</Product>
     <Description>Software framework for supporting coding best practices.</Description>
