+++ README.md
@@ -4,11 +4,12 @@ Framework for supporting coding best practices.
 
 ## Version control history
 
-**Oct 29, 2020 - 11.5.0 Resolved issue with UpdateAsync method of the Audit Decorator**
+**Oct 29, 2020 - 12.0.0 Resolved issue with UpdateAsync method of the Audit Decorator**
 
 - Resolved an issue with the UpdateAsync method of the Audit Decorator.
 - Redesigned the IRepository interface so that it no longer extends the IBulkRepository interface. Instead implementing classes need to implement IBulkRepository explicitly.
 - Renamed the Repository classes of the EntityFrameworkCore project to remove the "EntityFrameworkCore" prefix.
+- Renamed the Repository classes of the AzureStorage project to remove the "AzureStorage" prefix.
 - General code clean-up based on ReSharper recommendations.
 - Updated third-party NuGet packages.
